Fix wrong namespace references for unstable interfaces
A protocol as defined in "wlr-screencopy-unstable-v1.xml", named
"wlr_screencopy_unstable_v1" will emit something like this in
commons.zig

    pub const wlr = struct {pub usingnamespace @import("wlr_screencopy_unstable_v1_common.zig");};

under the name "wlr". That is the prefix derived from the protocol name.

Definitions in client/server however use the interface's name to get the
namespace, which for an interface such as "zwlr_screencopy_manager_v1"
would result in references to commons.zig like this:

    pub const Error = common.zwlr.layer_shell_v1.Error;

Those references break. This change passes the namespace from the
protocol down to the interface emitters.