Rework wrapping of wl_display functions
With the way libwayland uses errno, it isn't feasible for me to
determine exactly which errno values are possible after libwayland
function calls in most cases. Returning the full errno enum provided
by the zig standard and allowing the users to check for only those
values they care to handle differently is the most robust way to use
thesl APIs.

This change makes it so that the "number of events dispatched"
is no longer available, but I'm not aware of any use-case for this
information.