scanner: work around zig issue #131
It is possible, albeit a tad awkward, to work around #131 using @export
and @extern. The tricky part is ensuring that the corresponding @export
is always referenced/semantically analyzed if an @extern is referenced.

This is a breaking change.

Scanner.addCSource() is removed as it is unnecessary.

ObjectType.getInterface() is replaced by ObjectType.interface.

For example, wl.Seat.getInterface() becomes wl.Seat.interface.