build: make Scanner.create() more flexible/robust
Breaking changes:

1. The Scanner.Options fields are now LazyPaths which makes it possible
to provide the wayland-protocols installation through the zig package
manager for example. The field names have also been updated to remove
the redundant "_path" suffix.

2. The first argument to Scanner.create() is now the Build.Dependency
for zig-wayland. This allows us to get rid of the @src() hack currently
used to find the root source file for the scanner. This @src() hack will
stop working in Zig 0.19 so fixing this now prevents a second API break
in the near future.

Closes: https://codeberg.org/ifreund/zig-wayland/issues/66