build: do nothing in build() by default
To quote mitchellh's bug report/investigation:

The issue is in the *Build.Dependency argument to Scanner.create. When
b.dependency("x") is called, Zig executes its build function. The build
function for zig-wayland calls Scanner.create with default options,
forcing a pkg-config call.

Practically, this is affecting our downstream packagers of Ghostty to
adopt another dependency while we source the Wayland protocols XML files
from a build.zig.zon instead.

While a bit hacky, this fix is a simple way to fix the problem. Perhaps
std.Build will expose a better way to do this in the future.

Closes: https://codeberg.org/ifreund/zig-wayland/issues/67